# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Baby Health Application API
  version: 1.0.0
  description: Multi-baby health record API (India). Reference values are returned only with provenance and only when their release gate is CLEARED.
    The API never returns a computed milk target for healthy infants.
  contact:
    name: Product owner
servers:
- url: https://api.example.in
  description: Production (India region) - placeholder
security:
- bearerAuth: []
...
paths:
...
  /v1/references/sources:
    get:
      summary: List reference sources
      tags:
      - MedicalReferences
      x-authorization: Any authenticated user
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RefSource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/references/sources/{id}:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      summary: Get reference source
      tags:
      - MedicalReferences
      x-authorization: Any authenticated user
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefSource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/references/rules:
    get:
      summary: List CLEARED reference rules
      tags:
      - MedicalReferences
      x-authorization: Any authenticated user; non-CLEARED only with internal reviewer scope
      parameters:
      - name: population
        in: query
        schema:
          type: string
      - name: context
        in: query
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RefRule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/references/conflicts:
    get:
      summary: List source conflicts
      tags:
      - MedicalReferences
      x-authorization: Any authenticated user
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Conflict'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/references/unsupported:
    get:
      summary: List unsupported / not-established claims
      tags:
      - MedicalReferences
      x-authorization: Any authenticated user
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Unsupported'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
  /v1/babies/{babyId}/references/applicable:
    parameters:
    - $ref: '#/components/parameters/BabyId'
    - name: method
      in: query
      schema:
        enum:
        - DIRECT_BREASTFEEDING
        - EXPRESSED_BREASTMILK
        - FORMULA
        - OTHER
        - ANY
    get:
      summary: Resolve applicable reference for this baby now
      tags:
      - MedicalReferences
      x-authorization: Any active member of the baby (OWNER, GUARDIAN, CAREGIVER, VIEWER)
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
      description: Never returns a computed personal target. Healthy term => NO_AUTHORITATIVE_UNIVERSAL_VALUE_FOUND.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token <= 15 min; authorization resolved per baby via membership + database row-level security.
  parameters:
    BabyId:
      name: babyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resources of babies without membership return 404.
...
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/400
            title: Validation failed
            status: 400
    Unauthenticated:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/401
            title: Authentication required
            status: 401
    Forbidden:
      description: Role does not permit this action
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/403
            title: Role does not permit this action
            status: 403
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/404
            title: Not found
            status: 404
...
    RateLimited:
      description: Too many requests (see Retry-After)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/429
            title: Too many requests (see Retry-After)
            status: 429
    Internal:
      description: Internal error (support code provided)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://errors.example/500
            title: Internal error (support code provided)
            status: 500
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        support_code:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
    Audit:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        version:
          type: integer
    Provenance:
      type: object
      required:
      - source_id
      - organization
      - document_title
      - source_url
      - page_or_section
      - verification_status
      properties:
        source_id:
          type: string
        organization:
          type: string
        document_title:
          type: string
        document_type:
          type: string
        country:
          type: string
        population:
          type: string
        clinical_context:
          type: string
        source_url:
          type: string
        publication_date:
          type: string
        version:
          type: string
        page_or_section:
          type: string
        retrieved_at:
          type: string
        verified_at:
          type: string
        evidence_level:
          type: string
        verification_status:
          type: string
        notes:
          type: string
...
    FeedingPlanCreate:
      type: object
      required:
      - entered_from
      - clinician_name
      - instructed_on
      - plan_text
      - valid_from
      properties:
        entered_from:
          enum:
          - PRESCRIPTION
          - DISCHARGE_SUMMARY
          - VERBAL_INSTRUCTION
          - OTHER
        clinician_name:
          type: string
        instructed_on:
          type: string
          format: date
        plan_text:
          type: string
        volume_ml_per_feed:
          type:
          - number
          - 'null'
        feeds_per_day:
          type:
          - integer
          - 'null'
        valid_from:
          type: string
          format: date
        valid_to:
          type:
          - string
          - 'null'
          format: date
        attachment_document_id:
          type:
          - string
          - 'null'
          format: uuid
    FeedingPlan:
      allOf:
      - $ref: '#/components/schemas/FeedingPlanCreate'
      - $ref: '#/components/schemas/Audit'
      - type: object
        properties:
          id:
            type: string
            format: uuid
    ReferenceResult:
      type: object
      required:
      - reference_status
      properties:
        reference_status:
          enum:
          - NO_AUTHORITATIVE_UNIVERSAL_VALUE_FOUND
          - NOT_ESTABLISHED
          - CLINICAL_REFERENCE
          - CARE_TEAM_PLAN
          - CONFLICT_REQUIRES_REVIEW
          - REFERENCE_UNAVAILABLE_INVALID_PROFILE
        display_mode:
          type: string
        messages:
          type: array
          items:
            type: string
        rules:
          type: array
          items:
            type: object
            properties:
              rule_id:
                type: string
              population:
                type: string
              clinical_context:
                type: string
              value_min:
                type:
                - number
                - 'null'
              value_max:
                type:
                - number
                - 'null'
              value_target:
                type:
                - number
                - 'null'
              unit:
                type:
                - string
                - 'null'
              label:
                const: Clinical reference
              warnings:
                type: array
                items:
                  type: string
              provenance:
                $ref: '#/components/schemas/Provenance'
        plan:
          oneOf:
          - $ref: '#/components/schemas/FeedingPlan'
          - type: 'null'
        provenance:
          oneOf:
          - $ref: '#/components/schemas/Provenance'
          - type: 'null'
...
    RefSource:
      type: object
      properties:
        source_id:
          type: string
        organization:
          type: string
        document_title:
          type: string
        document_type:
          type: string
        country:
          type: string
        tier:
          type: string
        source_url:
          type: string
        publication_date:
          type: string
        version:
          type: string
        verification_status:
          type: string
        access_note:
          type: string
    RefRule:
      type: object
      properties:
        id:
          type: string
        population:
          type: string
        clinical_context:
          type: string
        release_gate:
          type: string
        automatic_calculation_allowed:
          type: boolean
        values:
          type: array
          items:
            type: object
        provenance:
          $ref: '#/components/schemas/Provenance'
    Conflict:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        source_a:
          type: string
        source_b:
          type: string
        difference:
          type: string
        population:
          type: string
        clinical_context:
          type: string
        date:
          type: string
        recommended_product_behavior:
          type: string
    Unsupported:
      type: object
      properties:
        id:
          type: string
        claim:
          type: string
        status:
          type: string
        product_behavior:
          type: string
...